Added user auth and also updated journal controller

+++ src/main/java/com/solution/journalapp/service/JournalEntryService.java
@@ -6,7 +6,6 @@ import com.solution.journalapp.repository.JournalEntryRepository;
 import lombok.extern.slf4j.Slf4j;
 import org.bson.types.ObjectId;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Component;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -28,10 +27,10 @@ public class JournalEntryService {
     public void saveEntry(JournalEntry journalEntry, String userName) {
         try{
             User user = userService.findByUserName(userName);
-            journalEntry.setDate(LocalDateTime.now());
+            journalEntry.setCreated_date(LocalDateTime.now());
             JournalEntry saved = journalEntryRepository.save(journalEntry);
             user.getJournalEntries().add(saved);
-            userService.saveEntry(user);
+            userService.saveUser(user);
         }catch (Exception e){
             log.error("Error saving journal entry", e);
         }
@@ -55,20 +54,27 @@ public class JournalEntryService {
         return journalEntryRepository.findById(id);
     }
 
-    public void deleteEntryById(ObjectId id, String userName){
+    @Transactional
+    public boolean deleteEntryById(ObjectId id, String userName){
+        boolean removed = false;
         try{
             User user = userService.findByUserName(userName);
-            user.getJournalEntries().removeIf(x->x.getId().equals(id));
-            userService.saveEntry(user);
-            journalEntryRepository.deleteById(id);
+            removed = user.getJournalEntries().removeIf(x -> x.getId().equals(id));
+            if(removed){
+                userService.saveUser(user);
+                journalEntryRepository.deleteById(id);
+            }
         }catch (Exception e){
+            System.out.println(e);
             log.error("Error saving journal entry", e);
+            throw new RuntimeException("An error occured while deleting this entry ",e);
         }
-
+        return removed;
 
     }
 
-    public void updateEntryById(ObjectId id, JournalEntry journalEntry){
+    public void updateEntryById(ObjectId id, JournalEntry journalEntry,String username){
+
         journalEntry.setId(id);
         journalEntryRepository.save(journalEntry);
     }
